tempReference/MXToAAF/requirements.txt
mutagen>=1.46
python-aaf2
